create the SubmittedForm component, fix button on this component, fix password fields properties
Create a separate SubmittedForm component, fix button link and position on this component.
Add helper text for the password field.
Remove password restriction by characters.
Add support for Cyrillic characters.
